+++ tests/testthat/test-lotri.R
@@ -105,3 +105,39 @@ test_that("default conditioning", {
                                                          .Dimnames = list(c("f", "g"), c("f", "g"))))))
 
 })
+
+test_that("zero diagonal etas must have zero off-diagonals (rxode2#481)", {
+  expect_error(
+    lotri({
+      lka <- 0.45
+      lcl <- 1
+      lvc <- 3.45
+      propSd <- fix(0)
+      etalcl + etalvc ~ c(0, 0.05, 0)
+    }),
+    regexp = "if diagonals are zero, off-diagonals must be zero for covariance matrices (row 1, column 2)",
+    fixed = TRUE
+  )
+  expect_error(
+    lotri({
+      lka <- 0.45
+      lcl <- 1
+      lvc <- 3.45
+      propSd <- fix(0)
+      etalcl + etalvc ~ c(1, 0.05, 0)
+    }),
+    regexp = "if diagonals are zero, off-diagonals must be zero for covariance matrices (row 2, column 1)",
+    fixed = TRUE
+  )
+  expect_error(
+    lotri({
+      lka <- 0.45
+      lcl <- 1
+      lvc <- 3.45
+      propSd <- fix(0)
+      etalcl + etalvc ~ c(0, 0.05, 1)
+    }),
+    regexp = "if diagonals are zero, off-diagonals must be zero for covariance matrices (row 1, column 2)",
+    fixed = TRUE
+  )
+})